+++ xdef/src/main/java/org/xdef/XDResultSet.java
@@ -49,7 +49,7 @@ public interface XDResultSet extends XDValue {
 	 */
 	public String itemAsString();
 
-	/** If the iterated object is a context then return relevant item value
+	/** If the iterated object is a Container then return relevant item value
 	 * as string; otherwise return null.
 	 * @param index the index of item.
 	 * @return value of the specified item as a string or return null.