+++ xdef/src/main/java/org/xdef/impl/ChkData.java
@@ -13,7 +13,6 @@ import org.xdef.XDValue;
 import org.xdef.XDValueAbstract;
 import org.xdef.component.XComponent;
 import org.xdef.impl.code.CodeUniqueset;
-import org.xdef.impl.code.DefParseResult;
 import org.xdef.impl.xml.KNamespace;
 import org.xdef.model.XMData;
 import org.xdef.model.XMDefinition;
@@ -35,7 +34,6 @@ import org.xdef.sys.SReporter;
 abstract class ChkData extends XDValueAbstract implements XXData {
 
 	final private XData _xDataModel;
-	private DefParseResult _parseResult;
 	final private ChkElement _parent;
 	private Object _userObject;
 	private final Node _node;
@@ -57,13 +55,11 @@ abstract class ChkData extends XDValueAbstract implements XXData {
 	 * @return The name of node.
 	 */
 	public final String getTextValue() {return _parent.getTextValue();}
-
 	@Override
 	/** Set text value to this node.
 	 * @param text the text value to be set.
 	 */
 	public final void setTextValue(String text) {_parent.setTextValue( text);}
-
 	@Override
 	/** Get parsed result of an attribute or text node.
 	 * @return parsed result of an attribute or text node.
@@ -71,19 +67,16 @@ abstract class ChkData extends XDValueAbstract implements XXData {
 	public final XDParseResult getParseResult() {
 		return _parent.getParseResult();
 	}
-
 	@Override
 	/** Get XDDocument.
 	 * @return XDDocument.
 	 */
 	public final XDDocument getXDDocument() {return _parent.getXDDocument();}
-
 	@Override
 	/** Get XDPool.
 	 * @return XDPool.
 	 */
 	public final XDPool getXDPool() {return _parent.getXDPool();}
-
 	@Override
 	/** Get root XXElement.
 	 * @return root XXElement node.
@@ -91,7 +84,6 @@ abstract class ChkData extends XDValueAbstract implements XXData {
 	public final XXElement getRootXXElement() {
 		return _parent.getRootXXElement();
 	}
-
 	@Override
 	/** Get actual associated XXElement.
 	 * @return root XXElement node.
@@ -120,7 +112,6 @@ abstract class ChkData extends XDValueAbstract implements XXData {
 	 * @return model of the processed data object.
 	 */
 	public final XMData getXMData() {return _xDataModel;}
-
 	@Override
 	/** Get namespace context of the parent element.
 	 * @return namespace context of the parent element.
@@ -128,21 +119,16 @@ abstract class ChkData extends XDValueAbstract implements XXData {
 	public final KNamespace getXXNamespaceContext() {
 		return _parent.getXXNamespaceContext();
 	}
-
 	@Override
 	public final String getXXName() {return _xDataModel.getName();}
-
 	@Override
 	public final String getXXNSURI() {return _xDataModel.getNSUri();}
-
 	@Override
 	public final Object getUserObject() {
 		return _userObject != null ? _userObject : _parent.getUserObject();
 	}
-
 	@Override
 	public final void setUserObject(Object obj) {_userObject = obj;}
-
 	@Override
 	/** Set named user object.
 	 * @param id identifier of the object.
@@ -152,7 +138,6 @@ abstract class ChkData extends XDValueAbstract implements XXData {
 	public final Object setUserObject(final String id, final Object obj) {
 		return _parent.setUserObject(id, obj);
 	}
-
 	@Override
 	/** Remove named user object.
 	 * @param id identifier of the object.
@@ -161,7 +146,6 @@ abstract class ChkData extends XDValueAbstract implements XXData {
 	public final Object removeUserObject(final String id) {
 		return _parent.removeUserObject(id);
 	}
-
 	@Override
 	/** Get named user object.
 	 * @param id identifier of the object.
@@ -170,176 +154,138 @@ abstract class ChkData extends XDValueAbstract implements XXData {
 	public final Object getUserObject(final String id) {
 		return _parent.getUserObject(id);
 	}
-
 	@Override
 	public final String getXPos() {return _parent.getXPos();}
-
 	@Override
 	/** Get source position.
 	 * @return source position or <i>null</i> if position is not available.
 	 */
 	public final SPosition getSPosition() {return _parent.getSPosition();}
-
 	@Override
 	public final XXNode getParent() {return _parent;}
-
 	@Override
 	public final XXNode[] getChildXXNodes() {return null;}
-
 	@Override
 	public final XMElement getXMElement() {return _parent.getXMElement();}
-
 	@Override
 	public final String getXDPosition() {return _parent.getXDPosition();}
-
 	@Override
 	public final XMDefinition getXMDefinition() {
 		return _parent.getXMDefinition();
 	}
-
 	@Override
 	public final String[] getVariableNames() {
 		return _parent.getVariableNames();
 	}
-
 	@Override
 	public final XDValue getVariable(String name) {
 		return _parent.getVariable(name);
 	}
-
 	@Override
 	public final void setVariable(final String name, final long value) {
 		_parent.setVariable(name, value);
 	}
-
 	@Override
 	public final void setVariable(final String name, final double value) {
 		_parent.setVariable(name, value);
 	}
-
 	@Override
 	public final void setVariable(final String name, final boolean value) {
 		_parent.setVariable(name, value);
 	}
-
 	@Override
 	public final void setVariable(final String name, final Object value) {
 		_parent.setVariable(name, value);
 	}
-
 	@Override
 	public final SReporter getReporter() {return _parent.getReporter();}
-
 	@Override
 	public final ReportWriter getReportWriter() {
 		return _parent.getReportWriter();
 	}
-
 	@Override
 	public final boolean errorWarnings() {return _parent.errorWarnings();}
-
 	@Override
 	public final boolean errors() {return _parent.errors();}
-
 	@Override
 	public final ArrayReporter getTemporaryReporter() {
 		return _parent.getTemporaryReporter();
 	}
-
 	@Override
 	public final ArrayReporter setTemporaryReporter(
 		final ArrayReporter reporter) {
 		return _parent.setTemporaryReporter(reporter);
 	}
-
 	@Override
 	public final void clearTemporaryReporter() {
 		_parent.clearTemporaryReporter();
 	}
-
 	@Override
 	public final boolean removeTemporaryReport(final Report rep) {
 		return _parent.removeTemporaryReport(rep);
 	}
-
 	@Override
 	public final boolean copyTemporaryReports() {
 		return _parent.copyTemporaryReports();
 	}
-
 	@Override
 	public final void putTemporaryReport(final Report message) {
 		_parent.putTemporaryReport(message);
 	}
-
 	@Override
 	public final boolean chkTemporaryErrors() {
 		return _parent.chkTemporaryErrors();
 	}
-
 	@Override
 	public final void fatal(final String id,
 		final String msg,
 		final Object... modif) {
 		_parent.fatal(id, msg, modif);
 	}
-
 	@Override
 	public final void error(final String id,
 		final String msg,
 		final Object... modif) {
 		_parent.error(id, msg, modif);
 	}
-
 	@Override
 	public final void warning(final String id,
 		final String msg,
 		final Object... modif) {
 		_parent.warning(id, msg, modif);
 	}
-
 	@Override
 	public final void fatal(final long id, final Object... modif) {
 		_parent.fatal(id, modif);
 	}
-
 	@Override
 	public final void error(final long id, final Object... modif) {
 		_parent.error(id, modif);
 	}
-
 	@Override
 	public final void warning(final long id, final Object... modif) {
 		_parent.warning(id, modif);
 	}
-
 	@Override
 	public final void putReport(final Report report){_parent.putReport(report);}
-
 	@Override
 	public final XPathFunctionResolver getXXFunctionResolver() {
 		return _parent.getXXFunctionResolver();
 	}
-
 	@Override
 	public final XPathVariableResolver getXXVariableResolver() {
 		return _parent.getXXVariableResolver();
 	}
-
 	@Override
 	public final XDValue getXDContext() {return _parent.getXDContext();}
-
 	@Override
 	public final void setXDContext(XDContainer xdc) {_parent.setXDContext(xdc);}
-
 	@Override
 	public final void setXDContext(final XDResultSet xdc) {
 		_parent.setXDContext(xdc);
 	}
-
 	@Override
 	public final void setXDContext(final Node node){_parent.setXDContext(node);}
-
 	@Override
 	/** Set source element as context for create mode.
 	 * @param source string with pathname, URL or source of XML node.
@@ -347,36 +293,28 @@ abstract class ChkData extends XDValueAbstract implements XXData {
 	public final void setXDContext(final String source) {
 		_parent.setXDContext(source);
 	}
-
 	@Override
 	public abstract short getItemId();
-
 	@Override
 	public final XDValue cloneItem() {return this;}
-
 	@Override
 	public final String stringValue() {
 		return _node == null ? null : _node.getNodeValue();
 	}
-
 	@Override
 	public final Node getXMLNode() {return _node;}
-
 	@Override
 	public final Element getElement() {return _parent.getElement();}
-
 	@Override
 	/** Get actual model.
 	 * @return actual model.
 	 */
 	public final XMNode getXMNode() {return _parent.getXMNode();}
-
 	@Override
 	/** Get XComponent.
 	 * @return The XComponent object (may be <i>null</i>).
 	 */
 	public final XComponent getXComponent() {return _parent.getXComponent();}
-
 	@Override
 	/** Set XComponent.
 	 * @param x XComponent object.
@@ -384,5 +322,4 @@ abstract class ChkData extends XDValueAbstract implements XXData {
 	public final void setXComponent(final XComponent x) {
 		_parent.setXComponent(x);
 	}
-
 }